Reduce remainder to uint8

+++ pkg/network/ebpf/c/protocols/kafka/types.h
@@ -52,7 +52,10 @@ typedef struct kafka_response_context_t {
     __s32 carry_over_offset;
     __u32 partitions_count;
     kafka_transaction_t transaction;
-    __u32 remainder;
+    // The number of remainder bytes stored from the previous packet into
+    // in remainder_buf. The maximum value is 3, even though remainder_buf
+    // needs to have space for 4 bytes to make building of the value easier.
+    __u8 remainder;
     char remainder_buf[4];
 } kafka_response_context_t;
